requirements.txt
openai
elevenlabs
moviepy==1.0.3
python-dotenv
requests
pydub
openai-whisper
